requirements.txt
Django>=6.0.3
google-auth>=2.0
google-auth-oauthlib>=1.0
google-auth-httplib2>=0.1
google-api-python-client>=2.0
python-dotenv>=1.0